+++ source/mysql/package.d
@@ -5,6 +5,7 @@ public import mysql.commands;
 public import mysql.connection;
 public import mysql.escape;
 public import mysql.exceptions;
+public import mysql.metadata;
 public import mysql.pool;
 public import mysql.prepared;
 public import mysql.protocol.constants;